requirements.txt
asttokens==2.4.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.0
executing==2.0.1
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pywin32==306
pyzmq==25.1.2
requests==2.32.3
soupsieve==2.6
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
urllib3==2.2.2
wcwidth==0.2.13
